Add in window centering with winit.

+++ swiftnav_console/backend_request_broker.py
@@ -63,11 +63,14 @@ class BackendRequestBroker(QObject):  # pylint: disable=too-many-instance-attrib
 
     @Slot()  # type: ignore
     def app_visible(self) -> None:  # pylint: disable=no-self-use
-        splash_filename = os.path.join(tempfile.gettempdir(), "splash_pid.txt")
-        try:
-            os.remove(splash_filename)
-        except FileNotFoundError:
-            pass
+        pid = os.getenv("SWIFT_CONSOLE_PID")
+        if pid is not None:
+
+            splash_filename = os.path.join(tempfile.gettempdir(), f"swiftnav_console.{pid}")
+            try:
+                os.remove(splash_filename)
+            except FileNotFoundError:
+                pass
 
     @Slot()  # type: ignore
     def serial_refresh(self) -> None:
